+++ component/frontend/Controller/Category.php
@@ -60,6 +60,7 @@ class Category extends DataController
 			'format'      => 'CMD',
 			'layout'      => 'CMD',
 			'id'          => 'INT',
+			'vgroupid'    => 'INT',
 		);
 
 		$urlparams = array_merge($additionalParams, $urlparams);