+++ component/frontend/View/Latest/Html.php
@@ -10,11 +10,10 @@ namespace Akeeba\ReleaseSystem\Site\View\Latest;
 defined('_JEXEC') or die;
 
 use Akeeba\ReleaseSystem\Site\Helper\Filter;
-use Akeeba\ReleaseSystem\Site\Helper\Title;
 use Akeeba\ReleaseSystem\Site\Model\Categories;
 use Akeeba\ReleaseSystem\Site\Model\Releases;
 use FOF30\Model\DataModel\Collection;
-use FOF30\View\View as BaseView;
+use FOF30\View\DataView\Html as BaseView;
 
 class Html extends BaseView
 {