+++ component/frontend/View/Releases/Html.php
@@ -11,13 +11,11 @@ defined('_JEXEC') or die;
 
 use Akeeba\ReleaseSystem\Site\Helper\Breadcrumbs;
 use Akeeba\ReleaseSystem\Site\Helper\Filter;
-use Akeeba\ReleaseSystem\Site\Helper\Router;
-use Akeeba\ReleaseSystem\Site\Helper\Title;
 use Akeeba\ReleaseSystem\Site\Model\Categories;
 use Akeeba\ReleaseSystem\Site\Model\Items;
 use Akeeba\ReleaseSystem\Site\Model\Releases;
 use FOF30\Model\DataModel\Collection;
-use FOF30\View\View as BaseView;
+use FOF30\View\DataView\Html as BaseView;
 
 class Html extends BaseView
 {
@@ -50,9 +48,6 @@ class Html extends BaseView
 
 	public function onBeforeBrowse($tpl = null)
 	{
-		// Prevent phpStorm's whining...
-		if ($tpl) {}
-
 		// Load the model
 		/** @var Releases $model */
 		$model = $this->getModel();