+++ component/frontend/View/Releases/tmpl/default.blade.php
@@ -21,7 +21,7 @@ defined('_JEXEC') or die;
 
 	@include('site:com_ars/Releases/category', ['id' => $this->category->id, 'item' => $this->category, 'Itemid' => $this->Itemid, 'no_link' => true])
 
-	<div class="ars-releases">
+	<div class="ars-releases ars-releases-{{ $this->category->is_supported ? 'supported' : 'unsupported' }}">
 	@if(count($this->items))
 		@foreach($this->items as $item)
 				@include('site:com_ars/Releases/release', ['item' => $item, 'Itemid' => $this->Itemid])